# Faster backwards for cumsum and cumprod