--- a/test/backward_compatibility/check_backward_compatibility.py
+++ b/test/backward_compatibility/check_backward_compatibility.py
@@ -79,6 +79,7 @@ allow_list = [
     ("aten::empty_meta", datetime.date(2021, 4, 1)),
     ("aten::batch_norm_backward_elemt", datetime.date(2021, 5, 1)),
     ("aten::assert_async", datetime.date(2021, 5, 1)),
+    ("aten::cumprod_backward", datetime.date(2021, 5, 1)),
 ]
 
 def allow_listed(schema, allow_list):