… aten/src/ATen/native/native_functions.yaml
… tools/autograd/derivatives.yaml
… torch/csrc/autograd/FunctionsManual.h