@nikitaved 2021-03-17
I have a question. cumprod is not implemented with the complex numbers support, then what about cumsum? Looks like now I have to redo my `prod` PR :)