@Lezcano 2021-03-17
@nikitaved 
**Complex support**. It does not have complex support, but I explain how to implement it at the end of the long comment. It basically accounts for doing:
```c++
const auto input_conj = input.conj();
const auto output_conj = output.conj();
```
and replace all the uses of `input` and `output`  by `input_conj` and `output_conj` all throughout the `cumprod_backward`.
**Backward of cumsum** It should not be changed, right? The complex derivative of cumsum is the same as the real derivative.

**gradgradcheck**. The modifications do pass `gradgradcheck`, as I left the original implementation for `gradgradcheck`. Are you getting fails? What's the fail? :/